Add a browsers file that exports envalid without fs access
