+++ browsers.d.ts
@@ -1,0 +1,2 @@
+// NOTE: dotEnvPath option will be ignored when using browsers.
+export * from './lib/envalid.d.ts'
