+++ lib/envalidWithDotenv.d.ts
@@ -1,0 +1,1 @@
+export * from './envalid.d.ts'
