+++ lib/extend.js
@@ -1,0 +1,3 @@
+const extend = (x = {}, y = {}) => Object.assign({}, x, y)
+
+module.exports = extend
