+++ package.json
@@ -3,6 +3,7 @@
   "version": "4.1.4",
   "description": "Validation for your environment variables",
   "main": "index.js",
+  "types": "index.d.ts",
   "files": [
     "index.js",
     "envalid.d.ts",
@@ -16,7 +17,7 @@
     "lint": "eslint . --ignore-pattern '!.eslintrc.js'",
     "lint-ts": "npm run prettier-ts -- --list-different",
     "test": "painless -r spec tests/*.js",
-    "prettier-ts": "prettier envalid.d.ts --tab-width 4 --print-width 100 --single-quote --no-semi",
+    "prettier-ts": "prettier lib/envalid.d.ts --tab-width 4 --print-width 100 --single-quote --no-semi",
     "coverage": "nyc npm test",
     "precommit": "lint-staged",
     "prepush": "npm run lint && npm run lint-ts && npm test"
